+++ esp-idf-svc/src/io.rs
@@ -1,2 +1,16 @@
 pub use embedded_svc::utils::io as utils;
 pub use esp_idf_hal::io::*;
+
+pub mod vfs {
+    use crate::sys;
+
+    #[allow(clippy::needless_update)]
+    pub fn initialize_eventfd(max_fds: usize) -> Result<(), sys::EspError> {
+        sys::esp!(unsafe {
+            sys::esp_vfs_eventfd_register(&sys::esp_vfs_eventfd_config_t {
+                max_fds: max_fds as _,
+                ..Default::default()
+            })
+        })
+    }
+}